requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
mysqlclient==2.2.7
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
jwt==1.3.1